pyproject.toml
[project]
name = "ml-dl-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "scikit-learn",
    "tensorflow",
    "pillow",
    "python-multipart",
    "spacy"
]

[dependency-groups]
dev = ["pytest>=8.0", "httpx>=0.27"]
